work on the holiday structure, date event emitter
