+++ .eslintrc.json
@@ -4,7 +4,7 @@
       "es2022": true,
       "node": true
   },
-  "extends": "standard",
+  "extends": "eslint:recommended",
   "overrides": [
   ],
   "parserOptions": {
